Move database models to separate ferrum-db crate
