+++ Cargo.toml
@@ -14,10 +14,12 @@ name = "ferrum"
 resolver = "2"
 members = [
     ".",
+    "ferrum-db",
     "ferrum-macros"
 ]
 
 [dependencies]
+ferrum-db = { path = "./ferrum-db" }
 actix = "0.12"
 actix-http = "3.0.0-beta.10"
 actix-web = "4.0.0-beta.9"
