+++ ferrum-db/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod users;
+pub mod servers;
+pub mod channels;
+pub mod messages;
