+++ ferrum-db/src/messages/mod.rs
@@ -1,0 +1,2 @@
+pub mod models;
+pub mod queries;
