+++ ferrum-db/src/users/mod.rs
@@ -1,0 +1,2 @@
+pub mod models;
+pub mod queries;
